+++ user-service/src/main/java/com/msyhu/naekaracubae/user/UserApplication.java
@@ -2,38 +2,13 @@ package com.msyhu.naekaracubae.user;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import springfox.documentation.builders.ApiInfoBuilder;
-import springfox.documentation.builders.PathSelectors;
-import springfox.documentation.builders.RequestHandlerSelectors;
-import springfox.documentation.service.ApiInfo;
-import springfox.documentation.spi.DocumentationType;
-import springfox.documentation.spring.web.plugins.Docket;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 
 @SpringBootApplication
-@EnableSwagger2
-@Configuration
 public class UserApplication {
 
     public static void main(String[] args) {
         SpringApplication.run(UserApplication.class, args);
-    }
-
-    @Bean
-    public Docket swaggerApi() {
-        return new Docket(DocumentationType.SWAGGER_2)
-                .select()
-                .apis(RequestHandlerSelectors.basePackage("com.msyhu.naekaracubae.user.resources"))
-                .paths(PathSelectors.any())
-                .build()
-                .apiInfo(new ApiInfoBuilder().version("1.0").title("User API").description("Documentation Employee API v1.0").build());
-    }
-
-    private ApiInfo apiInfo() {
-        return new ApiInfoBuilder().title("Web Application Api").description("New project Web Application Api ").version("1.0.0").license("Copyright blah blah").build();
     }
 
 }
