+++ user-service/src/main/java/com/msyhu/naekaracubae/user/config/SwaggerConfig.java
@@ -1,0 +1,30 @@
+package com.msyhu.naekaracubae.user.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.builders.PathSelectors;
+import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.spi.DocumentationType;
+import springfox.documentation.spring.web.plugins.Docket;
+import springfox.documentation.swagger2.annotations.EnableSwagger2;
+
+@EnableSwagger2
+@Configuration
+public class SwaggerConfig {
+
+    @Bean
+    public Docket swaggerApi() {
+        return new Docket(DocumentationType.SWAGGER_2)
+                .select()
+                .apis(RequestHandlerSelectors.basePackage("com.msyhu.naekaracubae.user.resources"))
+                .paths(PathSelectors.any())
+                .build()
+                .apiInfo(new ApiInfoBuilder().version("1.0").title("User API").description("Documentation Employee API v1.0").build());
+    }
+
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder().title("Web Application Api").description("New project Web Application Api ").version("1.0.0").license("Copyright blah blah").build();
+    }
+}
