+++ user-service/src/main/java/com/msyhu/naekaracubae/user/config/WebConfig.java
@@ -1,0 +1,16 @@
+package com.msyhu.naekaracubae.user.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class WebConfig implements WebMvcConfigurer {
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**")
+                .allowedOrigins("http://localhost:3000")
+                .allowedOrigins("http://naekaracubae.msyhu.com:80");
+    }
+}
